pyproject.toml
[tool.poetry]
authors = ["Alexander Regueiro <alexreg@gmail.com>"]
description = "Mirror mDNS hostnames on a DNS server"
name = "mdns-mirror"
packages = [{include = "mdns_mirror"}]
readme = "README.md"
version = "0.0.0"

[tool.poetry.dependencies]
dnspython = "^2.3.0"
typer_cloup = "^0.11.0"
zeroconf = "^0.47.3"

[tool.poetry.scripts]
mdns-mirror = "mdns_mirror.app:main"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
